Moved some files/dirs into shared folder
