+++ apps/account-recovery/.env
@@ -15,3 +15,4 @@
 # BASE_URL="https://safe-relay.rinkeby.gnosis.io/"
 # Mainnet staging
 BASE_URL="https://safe-relay.staging.gnosisdev.com/api"
+NETWORK="other" # | "rinkeby" | "mainnet"
