+++ apps/account-recovery/.env.production
@@ -1,0 +1,1 @@
+NETWORK="mainnet"
