+++ apps/account-recovery/App.tsx
@@ -2,15 +2,24 @@ import { StatusBar } from "expo-status-bar";
 import React from "react";
 import { SafeAreaProvider } from "react-native-safe-area-context";
 
-import useCachedResources from "./hooks/useCachedResources";
-import useColorScheme from "./hooks/useColorScheme";
-import Navigation from "./navigation";
+import useCachedResources from "./shared/hooks/useCachedResources";
+import useColorScheme from "./shared/hooks/useColorScheme";
+import Navigation from "./shared/navigation";
 
-import { AccountContextProvider } from "./context/AccountContext";
+import { AccountContextProvider } from "./shared/context/AccountContext";
+
+import Constants from "expo-constants";
 
 export default function App(): JSX.Element | null {
   const isLoadingComplete = useCachedResources();
   const colorScheme = useColorScheme();
+
+  const { manifest } = Constants;
+  const { extra } = manifest;
+  const { network } = extra;
+  console.log({ network });
+
+
 
   if (!isLoadingComplete) {
     return null;
