+++ apps/account-recovery/screens/NotFoundScreen.tsx
@@ -2,7 +2,7 @@ import { StackScreenProps } from "@react-navigation/stack";
 import * as React from "react";
 import { StyleSheet, Text, TouchableOpacity, View } from "react-native";
 
-import { RootStackParamList } from "../types";
+import { RootStackParamList } from "../shared/types";
 
 export default function NotFoundScreen({
   navigation,
