+++ apps/account-recovery/shared/navigation/BottomTabNavigator.tsx
@@ -3,10 +3,10 @@ import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import { createStackNavigator } from "@react-navigation/stack";
 import * as React from "react";
 
-import Colors from "../shared/constants/Colors";
+import Colors from "../constants/Colors";
 import useColorScheme from "../hooks/useColorScheme";
-import AccountScreen from "../screens/AccountScreen";
-import ContractBasedAccountScreen from "../screens/ContractBasedAccountScreen";
+import AccountScreen from "../../screens/AccountScreen";
+import ContractBasedAccountScreen from "../../screens/ContractBasedAccountScreen";
 import {
   BottomTabParamList,
   AccountParamList,
