+++ apps/account-recovery/shared/navigation/index.tsx
@@ -7,7 +7,7 @@ import { createStackNavigator } from "@react-navigation/stack";
 import * as React from "react";
 import { ColorSchemeName } from "react-native";
 
-import NotFoundScreen from "../screens/NotFoundScreen";
+import NotFoundScreen from "../../screens/NotFoundScreen";
 import { RootStackParamList } from "../types";
 import BottomTabNavigator from "./BottomTabNavigator";
 import LinkingConfiguration from "./LinkingConfiguration";
